+++ vcd/datasource_vcd_catalog_vapp_template.go
@@ -8,7 +8,8 @@ import (
 
 func datasourceVcdCatalogVappTemplate() *schema.Resource {
 	return &schema.Resource{
-		ReadContext: dataSourceVcdCatalogVappTemplateRead,
+		DeprecationMessage: "Use vcd_catalog_vapp_template instead",
+		ReadContext:        dataSourceVcdCatalogItemRead,
 		Schema: map[string]*schema.Schema{
 			"org": {
 				Type:     schema.TypeString,
@@ -24,7 +25,7 @@ func datasourceVcdCatalogVappTemplate() *schema.Resource {
 			"name": {
 				Type:         schema.TypeString,
 				Optional:     true,
-				Description:  "Name of the vApp Template. It is optional when a filter is provided",
+				Description:  "Name of the item. It is optional when a filter is provided",
 				ExactlyOneOf: []string{"name", "filter"},
 			},
 			"description": {
@@ -39,14 +40,19 @@ func datasourceVcdCatalogVappTemplate() *schema.Resource {
 			"metadata": {
 				Type:        schema.TypeMap,
 				Computed:    true,
-				Description: "Key and value pairs from the metadata of the vApp Template",
+				Description: "Key and value pairs from the metadata of the vApp template associated to this catalog item",
+			},
+			"catalog_item_metadata": {
+				Type:        schema.TypeMap,
+				Computed:    true,
+				Description: "Key and value pairs of the catalog item metadata",
 			},
 			"filter": {
 				Type:        schema.TypeList,
 				MaxItems:    1,
 				MinItems:    1,
 				Optional:    true,
-				Description: "Criteria for retrieving a vApp Template by various attributes",
+				Description: "Criteria for retrieving a catalog item by various attributes",
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"name_regex": elementNameRegex,
@@ -61,6 +67,6 @@ func datasourceVcdCatalogVappTemplate() *schema.Resource {
 	}
 }
 
-func dataSourceVcdCatalogVappTemplateRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	return genericVcdCatalogVappTemplateRead(ctx, d, meta, "datasource")
+func dataSourceVcdCatalogItemRead(_ context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	return genericVcdCatalogItemRead(d, meta, "datasource")
 }
